autopilot/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
supabase>=2.0.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
tqdm>=4.65.0
requests>=2.31.0
